Added StringValue Added YesFall (kinda... no) if something is broken blame the guy that commited the eclipse folder...

+++ src/minecraft/cat/util/lmao/MinecraftFontRenderer.java
@@ -19,7 +19,7 @@ public class MinecraftFontRenderer extends CFont {
 
     public static MinecraftFontRenderer createFontRenderer(Font font) {
         MinecraftFontRenderer c = new MinecraftFontRenderer(font, true, true);
-        c.FONT_HEIGHT = (int) (c.getHeight());
+        c.FONT_HEIGHT = c.getHeight();
         return c;
     }
